Add support for checksum on deploy and download
When deploy the checksum are calculated and sent
in headers to the server.
Deploy is first done using checksum only, no data
if the file doesn't exists on the server, the
exception is handled by trying to deploy again
with full data.

When downloading files, md5sum is calculated on
chunks after they are written to disk, when file
is closed the sum is compared to what Artifactory
has returned and an exception is raised when there
is a mismatch.

I didn't try again as in general this lib is not
doing retry on network issues, I assumed the calling
script is expected to handle that.